Fix defer, insecure
The previous commit made huge improvements in structure, regular Go formatting, etc. but unfortunately caused two regressions - defer resp.Body.Close() was called even when the connection fails (so it's nil and causes a crash) and the insecure flag was moved to a global var so the flag parsing needs to be updated to match. This commit fixes both of those problems.