Improve formating for contests, problems
Pretty print the contest and problem lists so that they're a little easier to read. Provided a good reason to add additional json attributes of both and add initial parsing and formatting of relative time for contest length. Next steps would be to order problems based on the ordinal, and parse and reformat the contest StartTime.

`Contests (2):
  finals: ICPC World Finals 2019
     5h starting at 2021-04-15T10:32:18.896-04
  comm: Practice contest
     2h starting at 2021-04-15T10:31:27.996-04`

`Problems (11):
  A: Azulejos
  B: Beautiful Bridges
  C: Checks Post Facto...`