Made object retrieval generic
The relevant API types have been extracted to a separate file and
interfaces have been created and implemented that enable retrieval of
data from the API. To aid with retrieval, I also implemented "id-based"
retrieval of data.

It might be a nice idea to extend this to accept a generic set of
parameters to enable retrieving of source code for instance.