Add judgements & judgement-types
We're going to need more contest types to make this interesting! Added judgement and judgement-type, so that we can tell which submissions have been judged and whether they're successful.

I've just added the types to the cli using the existing pattern and basic output for now. It's useful for testing, but I'd like to work toward human-friendly output like listing the current team's submissions and their judgements inline.