Fixed flag and double error issues
Due to Cobra passing flags around by reference instead of value
registering a flag as required resulted in undesired behavior. This was
fixed by checking the flags in the `RunPreE` callback for each command.
A heler has been introduced that generates a callback which checks the
flags and loads the flag-value into viper when it is set.

Before this change, when the command returns an error, it was printed.
Since Cobra also provides it own output, the error was logged twice. By
removing this second log statement, the "double error" issue was fixed.