Attempt at simplifying submission
I attempted to simplify the submit command. Though due to it "simple
complexity" it is difficult to actually make large changes that actually
increase readability. The only thing that could be done perhaps increase
readability, is to extract the separate "phases" of the command method
into seperate methods.

For example:
 - LoadAll()
 - DetectProblemAndLanguage()
 - DeduceEntryPoint()
 - Submit()

Though after having executed this refactor, it did not actually appear
to increase readability. So these changes have been reverted. But
another brave soul can attempt it if wanted.