Listing clarifications
This could use more work and maybe there's a better way to output clarifications, but for now it adds the initial command ("clar" for now) to list all of the team's clarifications. Each clarification is output with a message to say whether the team submitted it, it's a response, or a broadcast. The contest time is also shown, if it was associated with a problem, and the text, e.g.:

Clarifications (2):
  Clarification sent to jury at 1h38m58.21s for problem E (Dead-End Detector):
     In case n = 2 and m = 1, there are two solutions:
     1 2
     and
     2 1

     Is it allowed to print any of them?
  Response from jury at 1h39m54.235s for problem E (Dead-End Detector):
     > In case n = 2 and m = 1, there are two solutions:
     > 1 2
     > and
     > 2 1
     >
     > Is it allowed to print any of them?

     No comment