Default to the best contest
This change allows you to use most commands without specifying a contest id. If the API can connect and determine a "best" contest (e.g. if there is only one running contest or only one today) then it'll output a message and automatically connect to that one.

The algorithm is fairly simple and could likely be improved in the future, but it works for all the cases I could think of. This change covers the following commands: contest, clar, post-clar, and problem. Note that the submit command is *not* complete since it uses the config string directly - I'll submit another PR for that.