Submissions command
Adds a command to list all of a team's submissions, and their judgements, e.g.:
```./contest submissions
Automatically connecting to contest: ICPC World Finals 2019
Submissions (11):
  Submission to problem J: Miniature Golf in C++ at 34m14.979s
     Judged at 34m18.925s: WA (wrong answer)
  Submission to problem A: Azulejos in C++ at 1h1m4.046s
     Judged at 1h1m22.299s: AC (correct)
  Submission to problem E: Dead-End Detector in C++ at 1h41m34.267s
     Judged at 1h41m40.449s: WA (wrong answer)
```

Note this is *hard-coded* to team id "1" on line 77. The intention is that the role API being discussed at the CLICS meetings will provide a way to dynamically determine which team the current user belongs to.